refactor: clean up MUME tests and ensure compliance with robot rules
- Merged MUME Vitest tests into `tests/mume.test.js` and removed redundant test scripts.
- Replaced prohibited `robot.pow` automation scripts with compliant `mercenary.txt` management fixture.
- Migrated manual validation scripts to automated fixture test suite in `tests/fixtures/`.
- Removed `tests/manual_validation/` directory.
- Updated project memories to reflect ephemeral nature of root HTML documentation and MUME compliance rules.
