+++ tests/fixtures/aliases-manual.txt
