+++ tests/fixtures/xp_counter-manual.txt
