--- tests/manual_validation/robot.pow
@@ -1,17 +1,0 @@
-#print ("robot.pow") #(@auto_bash = 0) #(@auto_bless = 0)
-
-#alias bot+@robot={assist+;bash+;ride+;bless+} #alias bot-@robot={assist-;bash-;ride-;bless-}
-
-#alias bless+@robot={#(@auto_bless=1)} #alias bless-@robot={#(@auto_bless=0)}
-
-#action >+followleader@robot ^You now follow $1.={#print;#var $leader=$1}
-
-#alias assist+@robot={#(@assist_timer = timer);#action +botjoin1;#action +botjoin2;#action +botassist0;#action +botassist1;#action +botassist2;#action +botassist3;#action +botassist4;#action +botassist5;#action +botassist6;#action +botassist7;#action +botassist8;#action +botassist9;#action +botassist10;#action +botassist11} #alias assist-@robot={#var @assist_timer=;#action -botjoin1;#action -botjoin2;#action -botassist0;#action -botassist1;#action -botassist2;#action -botassist3;#action -botassist4;#action -botassist5;#action -botassist6;#action -botassist7;#action -botassist8;#action -botassist9;#action -botassist10;#action -botassist11} #action >-botjoin1@robot ^&1 joins &2'#&3fight.={#print;#(@-40=0);#($-41=($1:1));botjoinsub1;#if (@-40!=0) botjoinsub2;#else {#(@-40=0);#($-41=($2:1));botjoinsub1;#if (@-40!=0) botjoinsub2}} #action >-botjoin2@robot ^You assist &1={#print;#(@assist_timer=timer)} #action >+botjoin3@robot ^Your victim has disappeared!={#print;#(@assist_timer = timer - 3000)} #action >-botassist0@robot ^${leader}&1 approaches &2, trying to={#print;#(@assist_timer = timer - 3000);botjoinsub2} #action >-botassist1@robot ^${leader}&1 slashes={#print;botjoinsub2} #action >-botassist2@robot ^${leader}&1 cleaves={#print;botjoinsub2} #action >-botassist3@robot ^${leader}&1 smites={#print;botjoinsub2} #action >-botassist4@robot ^${leader}&1 hits={#print;botjoinsub2} #action >-botassist5@robot ^${leader}&1 pounds={#print;botjoinsub2} #action >-botassist6@robot ^${leader}&1 pierces={#print;botjoinsub2} #action >-botassist7@robot ^${leader}&1 stabs={#print;botjoinsub2} #action >-botassist8@robot ^${leader}&1 shoots={#print;botjoinsub2} #action >-botassist9@robot ^${leader}&1, but &2 parries successfully.={#print;botjoinsub2} #action >-botassist10@robot ^&1 swiftly dodges ${leader}&2 attempt to={#print;botjoinsub2} #action >-botassist11@robot ^\033[$1m${leader}&1 'assist'\033[0m={#print;botjoinsub2} #alias botjoinsub1@robot={#for (@1=1; @1<=(:?$-40); @1++) #if ($-40:@1==$-41) #(@-40=@1)} #alias botjoinsub2@robot={#if (.?$fig==0 && (((timer-@assist_timer)/1000) > 1)) {#(@assist_timer = timer);#send ("assist");botbashsub}}
-
-#alias bash+@robot={#action +botbash1;#action +botbash2;#(@auto_bash=1)} #alias bash-@robot={#action -botbash1;#action -botbash2;#(@auto_bash=0)} #action >-botbash1@robot ^&1 seems to have recovered &2 senses.={#print;#(@-40=0);#($-41=($1:1));botjoinsub1;#if (@-40==0) #send ("bash")} #action >-botbash2@robot avoids your bash, you topple over and lose your balance.={#print;#send ("bash")} #action >+botbash3@robot You intercept &1 blow.={#print;#(@assist_timer=timer);botbashsub} #alias botbashsub@robot={#if (@auto_bash==1) #send ("bash")} #action >+botbash4@robot Maybe learning how to bash would help?={#print;bash-}
-
-#alias reveal+@robot={#action +botreveal} #alias reveal-@robot={#action -botreveal} #action >-botreveal@robot ^${leader}&1 carefully examines the ground around him, looking for tracks...={#print;#send ("reveal quickly")}
-
-#alias ride+@robot={#action +zblam} #alias ride-@robot={#action -zblam} #action >+zblam@robot ^ZBLAM! &1doesn't want you riding={#print;stand;ride}
-
-#action >+merc1@robot A citizen mercenary ($1) taps you on the shoulder.={#print;#beep} #action >+merc2@robot A citizen mercenary taps you on the shoulder.={#print;#beep}
